LPS-93514 Move com.liferay.portal.util.* tests to the test module as Arquillian tests
