+++ modules/apps/portal/portal-util-test/src/testIntegration/java/com/liferay/portal/util/test/BasePortalImplURLTestCase.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.util;
+package com.liferay.portal.util.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.model.Group;
@@ -28,11 +29,13 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.util.test.LayoutTestUtil;
 
 import org.junit.Before;
+import org.junit.runner.RunWith;
 
 /**
  * @author Vilmos Papp
  * @author Akos Thurzo
  */
+@RunWith(Arquillian.class)
 public abstract class BasePortalImplURLTestCase {
 
 	@Before
