+++ modules/apps/portal/portal-util-test/src/testIntegration/java/com/liferay/portal/util/test/PortalImplLayoutFullURLTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.util;
+package com.liferay.portal.util.test;
 
+import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.portal.kernel.model.LayoutSet;
 import com.liferay.portal.kernel.service.VirtualHostLocalServiceUtil;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -26,10 +27,12 @@ import java.net.URL;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 
 /**
  * @author Akos Thurzo
  */
+@RunWith(Arquillian.class)
 public class PortalImplLayoutFullURLTest extends BasePortalImplURLTestCase {
 
 	@ClassRule
