+++ modules/apps/portal/portal-util-test/src/testIntegration/resources/com/liferay/portal/util/test/dependencies/test.jpg
